refactor(runner): split process discovery responsibilities
