+++ crates/runner/src/cmd/exec.rs
@@ -9,6 +9,7 @@ use sandbox::{SandboxControl, SandboxControlError};
 
 use crate::error::{RunnerError, RunnerResult};
 use crate::process;
+use crate::run_resolution;
 
 // ---------------------------------------------------------------------------
 // CLI args
@@ -99,8 +100,8 @@ pub async fn run_exec(args: ExecArgs, control: &dyn SandboxControl) -> RunnerRes
         sid.clone()
     } else if let Some(ref rid) = args.run {
         let discovered = process::discover_all().await;
-        let mappings = process::collect_active_run_mappings(&discovered.runners).await;
-        process::resolve_run_to_sandbox(rid, &mappings)?
+        let mappings = run_resolution::collect_active_run_mappings(&discovered.runners).await;
+        run_resolution::resolve_run_to_sandbox(rid, &mappings)?
     } else {
         // clap group guarantees one is set — this branch is unreachable.
         return Err(RunnerError::Config(
