+++ crates/runner/src/cmd/kill.rs
@@ -22,6 +22,7 @@ use tracing::info;
 
 use crate::error::{RunnerError, RunnerResult};
 use crate::process::{self, FirecrackerProcessInfo};
+use crate::run_resolution;
 
 // ---------------------------------------------------------------------------
 // CLI args
@@ -56,7 +57,7 @@ pub async fn run_kill(args: KillArgs, control: &dyn SandboxControl) -> RunnerRes
 
     // Phase 2: Resolve target — mode depends on which flag was passed.
     let target = if let Some(ref run_id) = args.run {
-        let mappings = process::collect_active_run_mappings(&discovered.runners).await;
+        let mappings = run_resolution::collect_active_run_mappings(&discovered.runners).await;
         resolve_by_run_id(run_id, &mappings, &discovered.firecrackers)?
     } else if let Some(ref sandbox_id) = args.sandbox {
         resolve_by_sandbox_id(sandbox_id, &discovered.firecrackers)?
@@ -125,14 +126,14 @@ pub async fn run_kill(args: KillArgs, control: &dyn SandboxControl) -> RunnerRes
 ///
 /// Maps run_id → sandbox_id via the provided mappings, then locates the FC
 /// by sandbox_id. The caller is responsible for collecting `mappings` via
-/// [`process::collect_active_run_mappings`] so this function stays pure and
+/// [`run_resolution::collect_active_run_mappings`] so this function stays pure and
 /// testable.
 fn resolve_by_run_id<'a>(
     input: &str,
-    mappings: &process::ActiveRunMappings,
+    mappings: &run_resolution::ActiveRunMappings,
     firecrackers: &'a [FirecrackerProcessInfo],
 ) -> RunnerResult<&'a FirecrackerProcessInfo> {
-    let sandbox_id = process::resolve_run_to_sandbox(input, mappings)?;
+    let sandbox_id = run_resolution::resolve_run_to_sandbox(input, mappings)?;
     firecrackers
         .iter()
         .find(|fc| fc.sandbox_id == sandbox_id)
@@ -301,131 +302,15 @@ mod tests {
         }
     }
 
-    // -- resolve_run_to_sandbox tests (shared helper in process.rs) ----------
-
     /// Build an `ActiveRunMappings` from a vec of `(run_id, sandbox_id)` pairs
     /// with zero read failures — the common test case.
-    fn mappings(entries: Vec<(String, String)>) -> process::ActiveRunMappings {
+    fn mappings(entries: Vec<(String, String)>) -> run_resolution::ActiveRunMappings {
         let total = if entries.is_empty() { 0 } else { 1 };
-        process::ActiveRunMappings {
+        run_resolution::ActiveRunMappings {
             entries,
             runners_total: total,
             runners_failed: 0,
         }
-    }
-
-    #[test]
-    fn run_prefix_resolves_to_sandbox_id() {
-        let status = mappings(vec![(
-            "550e8400-run-1111-2222-aaaaaaaaaaaa".into(),
-            "sbox-9999".into(),
-        )]);
-        let result = process::resolve_run_to_sandbox("550e8400", &status);
-        assert_eq!(result.unwrap(), "sbox-9999");
-    }
-
-    #[test]
-    fn run_prefix_full_uuid() {
-        let status = mappings(vec![(
-            "550e8400-e29b-41d4-a716-446655440000".into(),
-            "sbox-full".into(),
-        )]);
-        let result =
-            process::resolve_run_to_sandbox("550e8400-e29b-41d4-a716-446655440000", &status);
-        assert_eq!(result.unwrap(), "sbox-full");
-    }
-
-    #[test]
-    fn run_prefix_ambiguous() {
-        let status = mappings(vec![
-            ("abc-111".into(), "sbox-A".into()),
-            ("abc-222".into(), "sbox-B".into()),
-        ]);
-        let Err(e) = process::resolve_run_to_sandbox("abc", &status) else {
-            panic!("expected ambiguity error");
-        };
-        let msg = e.to_string();
-        assert!(msg.contains("ambiguous"), "{msg}");
-        assert!(msg.contains("abc-111"), "{msg}");
-        assert!(msg.contains("abc-222"), "{msg}");
-    }
-
-    #[test]
-    fn run_prefix_no_match() {
-        let status = mappings(vec![("abc-111".into(), "sbox-A".into())]);
-        let result = process::resolve_run_to_sandbox("deadbeef", &status);
-        assert!(result.is_err());
-    }
-
-    #[test]
-    fn run_prefix_empty_input() {
-        let empty = mappings(vec![]);
-        let result = process::resolve_run_to_sandbox("", &empty);
-        assert!(result.is_err());
-    }
-
-    #[test]
-    fn run_prefix_dedups_duplicate_entries() {
-        let status = mappings(vec![("R1".into(), "S1".into()), ("R1".into(), "S1".into())]);
-        let result = process::resolve_run_to_sandbox("R1", &status);
-        assert_eq!(result.unwrap(), "S1");
-    }
-
-    #[test]
-    fn run_prefix_dedup_preserves_true_ambiguity() {
-        let status = mappings(vec![
-            ("R1".into(), "S1".into()),
-            ("R1".into(), "S1".into()),
-            ("R2".into(), "S2".into()),
-        ]);
-        let Err(e) = process::resolve_run_to_sandbox("R", &status) else {
-            panic!("expected ambiguity");
-        };
-        let msg = e.to_string();
-        let r1_count = msg.matches("R1").count();
-        assert_eq!(r1_count, 1, "R1 should appear once after dedup: {msg}");
-        assert!(msg.contains("R2"), "{msg}");
-    }
-
-    #[test]
-    fn run_prefix_aggregated_across_runners() {
-        let status = mappings(vec![
-            ("aaa-111".into(), "sbox-A".into()),
-            ("bbb-222".into(), "sbox-B".into()),
-        ]);
-        assert_eq!(
-            process::resolve_run_to_sandbox("aaa", &status).unwrap(),
-            "sbox-A"
-        );
-        assert_eq!(
-            process::resolve_run_to_sandbox("bbb", &status).unwrap(),
-            "sbox-B"
-        );
-    }
-
-    #[test]
-    fn run_prefix_no_match_hints_unreadable_runners() {
-        let m = process::ActiveRunMappings {
-            entries: vec![],
-            runners_total: 3,
-            runners_failed: 2,
-        };
-        let err = process::resolve_run_to_sandbox("abc", &m).unwrap_err();
-        let msg = err.to_string();
-        assert!(msg.contains("2 of 3"), "{msg}");
-        assert!(msg.contains("unreadable"), "{msg}");
-    }
-
-    #[test]
-    fn run_prefix_no_match_hints_no_runners() {
-        let m = process::ActiveRunMappings {
-            entries: vec![],
-            runners_total: 0,
-            runners_failed: 0,
-        };
-        let err = process::resolve_run_to_sandbox("abc", &m).unwrap_err();
-        let msg = err.to_string();
-        assert!(msg.contains("no runner processes"), "{msg}");
     }
 
     // -- resolve_by_run_id (run_id → FC lookup via status + FC list) ---------
