+++ crates/runner/src/main.rs
@@ -28,6 +28,7 @@ mod proxy;
 mod r2_cache;
 mod resource_budget;
 mod retry;
+mod run_resolution;
 mod runner_dirname;
 mod runtime_overrides;
 mod status;
